+++ src/opentyr.h
@@ -27,7 +27,7 @@
 #include <stdlib.h>
 #include <string.h>
 
-#define COUNTOF(x) (sizeof(x) / sizeof *(x))  // use only on arrays!
+#define COUNTOF(x) ((unsigned)(sizeof(x) / sizeof *(x)))  // use only on arrays!
 #define MIN(a, b) ((a) < (b) ? (a) : (b))
 #define MAX(a, b) ((a) > (b) ? (a) : (b))
